--- a/models/tf.py
+++ b/models/tf.py
@@ -342,6 +342,9 @@ class TFProto(keras.layers.Layer):
         self.upsample = TFUpsample(None, scale_factor=2, mode='nearest')
         self.cv2 = TFConv(c_, c_, k=3, w=w.cv2)
         self.cv3 = TFConv(c_, c2, w=w.cv3)
+    
+    def call(self, inputs):
+        return self.cv2(self.cv2(self.upsample(self.cv1(inputs))))
 
 class TFUpsample(keras.layers.Layer):
     # TF version of torch.nn.Upsample()