--- a/models/yolo.py
+++ b/models/yolo.py
@@ -125,7 +125,7 @@ class DetectSegment(Detect):
             x[i] = x[i].view(bs, self.na, self.no, ny, nx).permute(0, 1, 3, 4, 2).contiguous()
 
             if not self.training:  # inference
-                if self.grid[i].shape[2:4] != x[i].shape[2:4] or self.onnx_dynamic:
+                if self.grid[i].shape[2:4] != x[i].shape[2:4] or self.dynamic:
                     self.grid[i], self.anchor_grid[i] = self._make_grid(nx, ny, i)
 
                 y = x[i].clone()