--- a/models/yolo.py
+++ b/models/yolo.py
@@ -231,8 +231,8 @@ class DetectionModel(BaseModel):
             s = 256  # 2x min stride
             m.inplace = self.inplace
             m.stride = torch.tensor([s / x.shape[-2] for x in self.forward(torch.zeros(1, ch, s, s))[0]])  # forward
-            m.anchors /= m.stride.view(-1, 1, 1)
             check_anchor_order(m)
+            m.anchors /= m.stride.view(-1, 1, 1)
             self.stride = m.stride
             self._initialize_biases()  # only run once
         elif isinstance(m, Detect):