… .github/workflows/ci-testing.yml
… data/coco128-seg.yaml
… data/coco128.yaml
… models/segment/yolov5l-seg.yaml
… models/segment/yolov5m-seg.yaml
… models/segment/yolov5n-seg.yaml
… models/segment/yolov5s-seg.yaml
… models/segment/yolov5x-seg.yaml
… requirements.txt
… segment/train.py
… segment/val.py
… utils/benchmarks.py
… utils/dataloaders.py
… utils/general.py
… utils/loggers/__init__.py
… utils/plots.py
… utils/segment/dataloaders.py
… utils/segment/general.py
… utils/segment/loss.py
… utils/segment/metrics.py
… utils/segment/plots.py
… val.py