Usage:
```
# train
python segment/train.py --cfg yolov5(m/s/l)-seg 
# detect
python segment/predict.py --source
```

TODO:
- [x] Resolve conflict
- [x] Resolve pre-commit 
- [x] Enable `overlap-mask` by default
- [x] Clean callback related residue code
- [ ] Write segmentation related readme
- [x] torch.deterministic solution for segment?
- [x] merge `non_max_suppression_masks()` with `non_max_suppression()`?
- [x] remove `edict` dependency?
- [x] DetectMultiBackend compatibility - PyTorch only
- [x] export.py compatability all non-TF
- [x] TensorBoard and W&B logging
- [x] CI - Test segmentation: Requires YOLOv5n-seg.pt and COCO128-seg dataset
- [x] export.py TF with TFDetectSegment()
- [x] Fix csv logging 
- [x] cpu support 
- [x] implement SegmentationModel class (requires retraining all models) to help identify seg models easily on load.
- [x] Fix Logging