@AyushExel 2022-08-20
@glenn-jocher I have sent you the trained weights files on slack. Once you add them to assets, we can update predict.py to get the weights for inference