@glenn-jocher 2022-08-20
@AyushExel precommit is all green now! FYI PyCharm with Code -> Reformat will catch 90% of these issues before they get to precommit. precommit can push fixes for some problems but not all of them.