@AyushExel 2022-08-20
@glenn-jocher ohh btw ignore segments/agumentations.py . All models are trained without albumentations installed so we'll probably end up removing that part from this PR.. before that I want to confirm if @Laughing-q has already made some progress on augmentation.