@glenn-jocher 2022-08-20
@AyushExel looking good. I added a few questions to the TODO list above. Also replied in Slack regarding weights.