@glenn-jocher 2022-08-20
> @glenn-jocher
> 
> > torch.deterministic solution for segment?
> 
> I think having train.py accept --deterministic should be the easiest way forward? segmentation uses interpolate with bilinear mode which is not deterministic on cuda.

resolved in https://github.com/ultralytics/yolov5/pull/9052/commits/8b70e64fca7a3b7d9f77012dc0371a7af72f9c78

> > merge non_max_suppression_masks() with non_max_suppression()?
> 
> I separated them just to future-proof the repo (was thinking of semantic seg, pose etc) as I don't know how similar or different NMS for other tasks will be. The current functions should be easy to merge. Its your call

yes we should definitely merge as much common code as we can to avoid future multiple-maintenance.

> > remove edict dependency?
> 
> Yeah we should remove that. I'll test with native dict and replace it

Perfect.