@AyushExel 2022-08-21
@glenn-jocher What should be the naming convention of the new models.. `yolov5s-seg.pt` or `yolov5s_seg.pt` ?
In the release assets both conventions are used. `yolov5-cls` and `effnet_**`.