@glenn-jocher 2022-08-21
> @glenn-jocher What should be the naming convention of the new models.. `yolov5s-seg.pt` or `yolov5s_seg.pt` ? In the release assets both conventions are used. `yolov5-cls` and `effnet_**`.

Good question. We already have the -cls models obviously so I'd just go with that. They pass CI on all OS's so I don't the dash is causing any problems. But I did look at the torchvision models and none of them have dashes, all underscores:

```
import torchvision
print(torchvision.models.__dict__)
```

> EDIT: @glenn-jocher I've added `output_to_target` with max filter limit. We need to limit number of predictions while plotting/inference otherwise it can cause cuda memory to run out or get the process stuck.

One option here is to pass all predictions to .cpu() before NMS. I think this only slows things down slightly.