@AyushExel 2022-08-21
@glenn-jocher Yes that masks are transferred to cpu but not limiting them to 15-20 masks per image causes the memory leak issue that we discussed few weeks ago. @Laughing-q and I can take a look at that after resolving other todos.

@Laughing-q we want to support coco format as well. If encountered, we'll just convert it to yolo-segments format and start training as usual. Do you want to incorporate your solution for overlapping labels here?