@glenn-jocher 2022-08-21
@AyushExel did some profiling in Colab with yolov5/val.py on COCO val. Looks like I was wrong, CPU NMS is much slower.

- 1.5 ms/img NMS default
- 7.9 ms/img to send model output to .cpu() before NMS
- 7.1 ms/img to send > conf_thresh outputs to .cpu() before NMS

Really strange, I'm pretty sure if I run on --device cpu the numbers aren't that bad. Let me try that on COCO128:

- 3.2ms/img for --device cpu

Ok so it seems like the GPU -> CPU transfer is playing a big part in the slowdown, and that native CPU NMS is about 2x slower than GPU (3.2ms vs 1.5ms)