@AyushExel 2022-08-21
@Laughing-q whenever you're back can you provide the following info:
* Is your fix for overlapping yolo-segments label ready to be merged? We want to support coco formats natively( see todos)
* Have you used albumentations in any experiment? I have benchmarked all models without that and so I was thinking of getting rid of that code.
* I have set the --mask-ratio to 4 be default. I think we should also overlap masks by default and allow to disable it by passing --no-overlap