@AyushExel 2022-08-21
@glenn-jocher @Laughing-q 
I think some changes in during the cleanup have broken a model in some way. I was just testing on s model and it gets significantly lower mAP in first 6 epochs.. But no worries it was working fine
 at the time this PR was made so we can just revert if nothing fixes it. 
I am double checking this. Will update here
<img width="348" alt="Screenshot 2022-08-21 at 11 10 19 PM" src="https://user-images.githubusercontent.com/15766192/185804107-f5391a06-d846-47f3-900c-693b7c759703.png">