@glenn-jocher 2022-08-21
@AyushExel interesting. This happened during classification and I tracked it down to doing EMA in FP16 instead of FP32, but it was hard to track down and I didn't realize the bug until I tried to train from the nearly ready to merge PR branch.