@glenn-jocher 2022-08-21
@AyushExel if the NMS was responsible then the train and val losses would be identical (but metrics would be lower). Is that the case?