@glenn-jocher 2022-08-21
@AyushExel I've unified NMS and uploaded yolov5s-seg.pt to assets, tested with segment/predict and all good. I've had a chance to review the PR more fully. I think I can be most useful in bringing the 3 files in yolov5/segment up to master so that they are as nearly identical to their detect counterparts as possible, including loading via DetectMultiBackend.