@glenn-jocher 2022-08-21
@AyushExel I've implemented updates to DetectMultiBackend to make it compatible with the PyTorch segmentation models and have made a big update to segment/predict.py to make it nearly equal to detect.py. Tested and works well. I'll try val next, maybe tomorrow, and then finally train.py later this week.