@glenn-jocher 2022-08-21
@AyushExel BTW can we uncomment the 3 generator lines here? These are supposed to assist with reproducibility and should not have an impact on the torch non-deterministic layers.
<img width="1292" alt="Screenshot 2022-08-22 at 01 52 14" src="https://user-images.githubusercontent.com/26833433/185816371-ce3f8112-4e58-4d5a-82c5-9807aa5b8146.png">