@Laughing-q 2022-08-22
@AyushExel 
> * Is your fix for overlapping yolo-segments label ready to be merged? We want to support coco formats natively( see todos)

you mean to add the fix to this segmentation repo or [ultralytics/JSON2YOLO](https://github.com/ultralytics/JSON2YOLO)?

> * Have you used albumentations in any experiment? I have benchmarked all models without that and so I was thinking of getting rid of that code.

I have run a experiment with this last night, the metric curves look almost the same with or without albumentations. [dashboard](https://wandb.ai/laughing/benchmark?workspace=user-laughing) `yolov5s-mr4-overlap-albu` vs `yolov5s-mr4-overlap`.

> * I have set the --mask-ratio to 4 be default. I think we should also overlap masks by default and allow to disable it by passing --no-overlap

yes I think so.