@glenn-jocher 2022-08-25
@AyushExel @Laughing-q I've finished updating val.py process_batch in https://github.com/ultralytics/yolov5/pull/9052/commits/252b8b32078e8f761385238813971e00653ffe93. I unified mask and box functions into 1 to reduce code duplication and resolved mAP bug on masks. New mask mAP is slightly higher, no speed impact (pycocotools mAP should stay the same, haven't tested).

Before
```
YOLOv5n-seg summary: 221 layers, 1986637 parameters, 0 gradients, 7.1 GFLOPs
val: Scanning '/Users/glennjocher/PycharmProjects/datasets/coco128-seg/labels/train2017.cache' images and labels... 126 found, 2 missing, 0 empty, 0 corrupt: 100%|██████████| 128/128 [00:00<?, ?it/s]
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95)     Mask(P          R      mAP50  mAP50-95): 100%|██████████| 4/4 [00:08<00:00,  2.22s/it]
                   all        128        929      0.849      0.603      0.688      0.477      0.714      0.493      0.534      0.279
Speed: 0.3ms pre-process, 48.1ms inference, 6.1ms NMS per image at shape (32, 3, 320, 320)
dt=11.092143058776855
```

NEW
```
YOLOv5n-seg summary: 221 layers, 1986637 parameters, 0 gradients, 7.1 GFLOPs
val: Scanning '/Users/glennjocher/PycharmProjects/datasets/coco128-seg/labels/train2017.cache' images and labels... 126 found, 2 missing, 0 empty, 0 corrupt: 100%|██████████| 128/128 [00:00<?, ?it/s]
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95)     Mask(P          R      mAP50  mAP50-95): 100%|██████████| 4/4 [00:08<00:00,  2.23s/it]
                   all        128        929      0.849      0.603      0.688      0.477      0.714      0.493      0.534      0.283
Speed: 0.2ms pre-process, 47.6ms inference, 5.9ms NMS per image at shape (32, 3, 320, 320)
Results saved to ../runs/val-seg/exp17
dt=11.088944911956787
```